Book CRUD methods

+++ src/main/java/com/vermeg/bookstore/controllers/BooksRestController.java
@@ -1,5 +1,6 @@
 package com.vermeg.bookstore.controllers;
 
+import com.vermeg.bookstore.Exceptions.BookNotFoundException;
 import com.vermeg.bookstore.entities.Book;
 import com.vermeg.bookstore.repositories.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +9,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/bookstore/")
@@ -22,31 +24,40 @@ public class BooksRestController {
     }
 
     @GetMapping("book/{id}")
-    public Book getBookById(@PathVariable int id){
-        return bookRepository.findById(id).orElseThrow(()-> new IllegalArgumentException());
+    public Book getBookById(@PathVariable int id) throws BookNotFoundException {
+        return bookRepository.findById(id).orElseThrow(()-> new BookNotFoundException("The book " +
+                "with the ID "+id+" does not exist"));
     }
 
     @PostMapping("book/add")
-    public void addBook(@RequestBody @Validated Book b,  BindingResult result) {
+    public Book addBook(@RequestBody @Validated Book b,  BindingResult result) {
         if(result.hasErrors())
             System.err.println(result.getAllErrors());
         // si le produit est valide
-        bookRepository.save(b);
-        System.out.println("Book added successfully");
+        return bookRepository.save(b);
     }
 
     @DeleteMapping("book/{id}")
-    public void deleteBook(@PathVariable int id){
-        bookRepository.deleteById(id);
+    public Optional<Book> deleteBook(@PathVariable int id) throws BookNotFoundException {
+        Optional<Book> b = bookRepository.findById(id);
+        if (b != null) {
+            bookRepository.deleteById(id);
+            return b;
+        }
+        else throw new BookNotFoundException("The book " +
+                "with the ID "+id+" does not exist");
     }
 
     @PutMapping("book/modify/{id}")
-    public void updateBook(@RequestBody @Validated Book b, @PathVariable int Id,
-                           BindingResult result) {
+    public void updateBook(@RequestBody @Validated Book b, @PathVariable int id,
+                           BindingResult result) throws BookNotFoundException {
         if(result.hasErrors())
             System.err.println(result.getAllErrors());
         // si le produit est valide
-        bookRepository.save(b);
-        System.out.println("Book update successfully");
+       bookRepository.findById(id).map(book -> bookRepository.save(b))
+               .orElseThrow(()-> new BookNotFoundException("The book " +
+               "with the ID "+id+" does not exist"));
     }
+
+
 }
